Add files via upload
The gradient operator in image processing is used to detect edges by highlighting areas where the intensity of an image changes rapidly. The operator computes the rate of change in pixel intensity in both the horizontal and vertical directions. It is commonly used in edge detection tasks, such as detecting boundaries, shapes, and transitions between objects in an image.

exp 4b Robert operator The Roberts operator is a simple edge detection operator used in image processing. It is designed to compute the gradient of the image at each pixel, highlighting regions where there is a rapid intensity change, often indicating an edge.

The Roberts operator uses a pair of 2x2 convolution kernels to calculate the gradients in the horizontal and vertical directions. The small kernel size (2x2) makes it more sensitive to noise compared to other operators like the Sobel operator, but it can be effective for detecting fine edges in an image.